fix: resolve merge conflicts and split community pages
- Fix unresolved merge artifacts in QuizPage, MyWordbookPage, CommunityPage, Navbar
- Split CommunityPage into CommunityPage (N05 wordbook share) and PostsPage (N04 board)
- Add /posts route and Navbar entry for community board
