+++ edulex/src/components/Navbar.jsx
@@ -1,7 +1,6 @@
 import { useState, useRef, useEffect } from 'react'
 import { Link, useLocation, useNavigate } from 'react-router-dom'
-import { Library, NotebookText, FlaskConical, BarChart2, LogOut, Users } from 'lucide-react'
-import { Library, NotebookText, FlaskConical, BarChart2, LogOut, Users, Trophy, MessageSquare, KeyRound } from 'lucide-react'
+import { Library, NotebookText, FlaskConical, BarChart2, LogOut, Users, Trophy, MessageSquare, KeyRound, Newspaper } from 'lucide-react'
 import { useAuth } from '../context/AuthContext'
 import { LIB } from '../constants/theme'
 import logo from '../assets/logo.png'
@@ -12,7 +11,8 @@ const MENU_ITEMS = [
   { path: '/wordbook/my',       label: '나만의 단어장', icon: <NotebookText size={18} strokeWidth={1.8} /> },
   { path: '/quiz',              label: '테스트',       icon: <FlaskConical size={18} strokeWidth={1.8} /> },
   { path: '/dashboard',        label: '학습 현황',    icon: <BarChart2 size={18} strokeWidth={1.8} /> },
-  { path: '/community',        label: '커뮤니티',     icon: <Users size={18} strokeWidth={1.8} /> },
+  { path: '/community',        label: '단어장 공유',   icon: <Users size={18} strokeWidth={1.8} /> },
+  { path: '/posts',            label: '커뮤니티',     icon: <Newspaper size={18} strokeWidth={1.8} /> },
   { path: '/ranking',          label: '랭킹',         icon: <Trophy size={18} strokeWidth={1.8} /> },
   { path: '/suggestions',      label: '건의사항',     icon: <MessageSquare size={18} strokeWidth={1.8} /> },
 ]
