+++ edulex/src/pages/QuizPage.jsx
@@ -1,17 +1,15 @@
-
 import { useEffect, useRef, useState } from 'react'
 import {
   BookOpen, Bot, Lock, Check, Bookmark, Trophy, BookMarked, Dumbbell, RotateCcw,
   ChevronDown, ChevronUp, TrendingUp, RefreshCw,
 } from 'lucide-react'
-
 import { supabase } from '../utils/supabase'
 import { useAuth } from '../context/AuthContext'
 import { useReward } from '../context/RewardContext'
 import { LIB, BOOK_COLORS } from '../constants/theme'
 import { useWordbookAchievement } from '../hooks/useWordbookAchievement'
-import { QUIZ_BOOK_ANIM_ID } from '../constants/character'
 import { useStudyTime } from '../hooks/useStudyTime'
+import { QUIZ_BOOK_ANIM_ID } from '../constants/character'
 
 
 // ── 유틸 ──────────────────────────────────────────────────────────
@@ -25,35 +23,6 @@ function shuffle(arr) {
   return a
 }
 
- feat/quiz-level
-function buildQuestions(targetWords, allWords, level = 1) {
-  const pool = shuffle(targetWords).slice(0, Math.min(10, targetWords.length))
-  return pool.map(word => {
-    let question, answer, wrongs
-    let choices = []
-    if (level === 1) {
-      question = word.major_meaning
-      answer = word.english
-      wrongs = shuffle(allWords.filter(w => w.id !== word.id)).slice(0, 3).map(w => w.english)
-      choices = shuffle([answer, ...wrongs])
-    } else if (level === 2) {
-      question = word.english
-      answer = word.major_meaning
-      wrongs = shuffle(allWords.filter(w => w.id !== word.id)).slice(0, 3).map(w => w.major_meaning)
-      choices = shuffle([answer, ...wrongs])
-    } else if (level === 3) {
-      question = word.major_meaning
-      answer = word.english
-    } else if (level === 4) {
-      const example = word.major_example || word.general_example || ''
-      const regex = new RegExp(word.english, 'gi')
-      question = example.replace(regex, '_________')
-      if (question === example) {
-        question = `(예문 없음) 뜻: ${word.major_meaning}`
-      }
-      answer = word.english
-    }
-    return { id: word.id, question, answer, choices }
 function buildQuestions(pool, source) {
   return shuffle(pool).map(word => {
     const wrongs  = shuffle(source.filter(w => w.id !== word.id)).slice(0, 3).map(w => w.english)
@@ -65,11 +34,6 @@ function buildQuestions(pool, source) {
 const CHOICE_LABELS = ['A', 'B', 'C', 'D']
 
 const LEVELS = [
- feat/quiz-level
-  { lv: 1, label: 'Lv 1', desc: '뜻 → 영어 객관식', locked: false },
-  { lv: 2, label: 'Lv 2', desc: '영어 → 뜻 객관식', locked: false },
-  { lv: 3, label: 'Lv 3', desc: '뜻 → 영어 주관식', locked: false },
-  { lv: 4, label: 'Lv 4', desc: '영어 예문 빈칸채우기', locked: false },
   { lv: 1, label: 'Lv 1', desc: '뜻 → 영어 객관식' },
   { lv: 2, label: 'Lv 2', desc: '영어 → 뜻 객관식' },
   { lv: 3, label: 'Lv 3', desc: '뜻 → 영어 주관식' },
@@ -104,14 +68,6 @@ function NoWordsModal({ level, onClose }) {
 }
 
 // ── 선택 화면 ─────────────────────────────────────────────────────
-
-function SelectView({ wordbooks, selectedWb, onSelect, onStart, levelStats, loadingWb }) {
-  const [activeLevel, setActiveLevel] = useState(1)
-  const [wbTab, setWbTab] = useState('official') // 'official' | 'user'
-
-  const officialBooks = wordbooks.filter(w => w.type === 'official')
-  const userBooks = wordbooks.filter(w => w.type === 'user')
-  const visibleBooks = wbTab === 'official' ? officialBooks : userBooks
 
 function SelectView({ wordbooks, selectedWb, onSelect, onStart, activeLevel, onLevelChange, progress, isReviewMode, wbLevelCounts, wordCountMode, onWordCountModeChange }) {
   const [wbTab, setWbTab]           = useState('official')
@@ -150,12 +106,6 @@ function SelectView({ wordbooks, selectedWb, onSelect, onStart, activeLevel, onL
             {LEVELS.map(({ lv, label, desc }) => {
               const locked   = false
               const isActive = activeLevel === lv
-              const isLocked = selectedWb ? (levelStats[lv] || 0) === 0 : true
-              return (
-                <button
-                  key={lv}
-                  onClick={() => !isLocked && setActiveLevel(lv)}
-                  disabled={isLocked}
               const count    = lvCounts ? lvCounts[lv - 1] : null
 
               return (
@@ -164,31 +114,13 @@ function SelectView({ wordbooks, selectedWb, onSelect, onStart, activeLevel, onL
                   onClick={() => !locked && onLevelChange(lv)}
                   disabled={locked}
                   className="relative flex flex-col items-center gap-1.5 py-4 rounded-xl border-2 transition-all duration-150"
-                  style={isLocked
+                  style={locked
                     ? { background: LIB.parchmentDark, borderColor: 'transparent', cursor: 'not-allowed', opacity: 0.55 }
                     : isActive
                       ? { background: `linear-gradient(135deg, ${LIB.wood} 0%, ${LIB.woodLight} 100%)`, borderColor: LIB.gold, boxShadow: `0 4px 14px rgba(92,58,30,0.3)` }
                       : { background: 'white', borderColor: LIB.shelfLine }
                   }
                 >
-                  <span className="absolute top-2 right-2 flex flex-col items-end gap-1">
-                    {isLocked && <Lock size={11} strokeWidth={2} style={{ color: LIB.shelfLine }} />}
-                    {!isLocked && selectedWb && (
-                      <span className="text-[9px] font-bold px-1.5 py-0.5 rounded-full" style={{ background: isActive ? LIB.goldLight : LIB.parchmentDark, color: isActive ? LIB.wood : LIB.inkMid }}>
-                        {levelStats[lv] || 0}
-                      </span>
-                    )}
-                  </span>
-                  <span
-                    className="text-sm font-black"
-                    style={{ color: isLocked ? LIB.shelfLine : isActive ? LIB.gold : LIB.ink }}
-                  >
-                    {label}
-                  </span>
-                  <span
-                    className="text-[9px] font-semibold text-center leading-tight px-1"
-                    style={{ color: isLocked ? LIB.shelfLine : isActive ? LIB.goldLight : LIB.inkLight }}
-                  >
                   {locked && (
                     <span className="absolute top-2 right-2">
                       <Lock size={11} strokeWidth={2} style={{ color: LIB.shelfLine }} />
@@ -276,7 +208,7 @@ function SelectView({ wordbooks, selectedWb, onSelect, onStart, activeLevel, onL
           <div className="flex rounded-xl p-1 mb-3 gap-1" style={{ background: LIB.parchmentDark }}>
             {[
               { key: 'official', icon: <BookOpen size={13} strokeWidth={2} />, label: '공식 단어장', count: officialBooks.length },
-              { key: 'user', icon: <Bot size={13} strokeWidth={2} />, label: '나만의 단어장', count: userBooks.length },
+              { key: 'user',     icon: <Bot size={13} strokeWidth={2} />,      label: '나만의 단어장', count: userBooks.length },
             ].map(tab => (
               <button
                 key={tab.key}
@@ -307,10 +239,6 @@ function SelectView({ wordbooks, selectedWb, onSelect, onStart, activeLevel, onL
                 {wbTab === 'user' ? '아직 나만의 단어장이 없어요' : '공식 단어장이 없어요'}
               </div>
             ) : visibleBooks.map((wb, idx) => {
-feat/quiz-level
-              const bookCol = BOOK_COLORS[idx % BOOK_COLORS.length]
-              const selected = selectedWb?.id === wb.id
-
               const bookCol    = BOOK_COLORS[idx % BOOK_COLORS.length]
               const selected   = selectedWb?.id === wb.id
               const lvCount    = wbLevelCounts?.[wb.id]?.[activeLevel] ?? null
@@ -384,8 +312,8 @@ feat/quiz-level
 
         {/* 시작 버튼 */}
         <button
-          onClick={() => onStart(activeLevel)}
-          disabled={!selectedWb || loadingWb || (levelStats[activeLevel] || 0) === 0}
+          onClick={onStart}
+          disabled={!selectedWb}
           className="w-full py-4 rounded-xl text-sm font-black tracking-wide transition-all duration-200 disabled:opacity-40"
           style={{
             background: selectedWb
@@ -395,7 +323,7 @@ feat/quiz-level
             boxShadow: selectedWb ? '0 4px 14px rgba(92,58,30,0.35)' : 'none',
           }}
         >
-          {loadingWb ? '단어장 불러오는 중...' : selectedWb ? `Lv ${activeLevel} 테스트 시작 →` : '단어장을 선택하세요'}
+          {selectedWb ? `Lv ${activeLevel} 테스트 시작 →` : '단어장을 선택하세요'}
         </button>
 
       </div>
@@ -403,48 +331,8 @@ feat/quiz-level
   )
 }
 
-// ── 퀴즈 화면 ─────────────────────────────────────────────────────
-
- feat/quiz-level
-function QuizView({ question, current, total, chosen, onChoose, level }) {
-  const [inputValue, setInputValue] = useState('')
-  const isSubjective = level >= 3
-
-  useEffect(() => {
-    setInputValue('')
-  }, [current])
-
-  const handleInputSubmit = (e) => {
-    e.preventDefault()
-    if (chosen !== null || !inputValue.trim()) return
-    onChoose(inputValue.trim().toLowerCase())
-  }
-
-  let titleText = '이 뜻에 해당하는 영어 단어는?'
-  if (level === 2) titleText = '이 단어의 알맞은 뜻은?'
-  if (level === 3) titleText = '이 뜻에 해당하는 영어 단어를 입력하세요'
-  if (level === 4) titleText = '빈칸에 들어갈 알맞은 영어 단어는?'
-
-  let inputColor = LIB.ink
-  let inputBg = 'white'
-  let inputBorder = LIB.shelfLine
-  if (chosen !== null && isSubjective) {
-    const isCorrect = chosen.trim().toLowerCase() === question.answer.toLowerCase()
-    if (isCorrect) {
-      inputBg = '#f0fdf4'
-      inputBorder = '#4ade80'
-      inputColor = '#166534'
-    } else {
-      inputBg = '#fff1f2'
-      inputBorder = '#f87171'
-      inputColor = '#991b1b'
-    }
-  }
-
-
 // ── 책 낙하/쌓기 애니메이션 (Q02) ────────────────────────────────
 
-// 폭발 방향 벡터 — 하단 기준으로 방사형 (index 0=하단 책부터 먼저 폭발)
 const EXPLODE_VECTORS = [
   { tx: '-80vw', ty: '-15vh', ts: 2.5, tr: '-130deg' },
   { tx: '-55vw', ty: '-65vh', ts: 2,   tr:   '90deg' },
@@ -486,7 +374,6 @@ function BookStackAnimation({ count, isToppling, total }) {
               borderRadius: 3,
               background: `linear-gradient(to right, ${c.spine} 0%, ${c.spine} 7%, ${c.cover} 7%, ${c.cover} 93%, ${c.accent} 93%)`,
               boxShadow: '0 4px 12px rgba(0,0,0,0.3)',
-              // 아래 책(i=0)이 먼저 폭발, 위 책이 약간 늦게
               animationDelay: isToppling ? `${i * 25}ms` : '0ms',
               '--tx': v.tx, '--ty': v.ty, '--ts': v.ts, '--tr': v.tr,
             }}
@@ -528,9 +415,6 @@ function QuizView({ question, current, total, chosen, onChoose, activeLevel, has
             ))}
           </div>
           <span className="text-[10px] font-bold px-2 py-0.5 rounded-full shrink-0" style={{ background: LIB.parchmentDark, color: LIB.inkMid }}>
- feat/quiz-level
-            Lv {level}
-
             Lv {activeLevel}
           </span>
         </div>
@@ -540,53 +424,14 @@ function QuizView({ question, current, total, chosen, onChoose, activeLevel, has
           <div className="absolute top-0 bottom-0 left-1/2 -translate-x-px w-px opacity-15" style={{ background: LIB.shelfLine }} />
           <div className="absolute top-0 right-8 w-3 h-8 rounded-b-sm opacity-60" style={{ background: LIB.deepRed }} />
           <p className="text-xs font-semibold mb-3 uppercase tracking-widest" style={{ color: LIB.inkLight }}>
-            {titleText}
+            이 뜻에 해당하는 영어 단어는?
           </p>
           <p className="text-2xl font-black leading-snug" style={{ color: LIB.ink }}>{question.question}</p>
         </div>
 
- feat/quiz-level
-        {/* 선택지 또는 입력 폼 */}
-        {isSubjective ? (
-          <form onSubmit={handleInputSubmit} className="flex flex-col gap-3">
-            <input
-              type="text"
-              value={inputValue}
-              onChange={e => setInputValue(e.target.value)}
-              disabled={chosen !== null}
-              placeholder="정답을 입력하세요"
-              autoFocus
-              className="w-full px-5 py-4 rounded-xl border-2 text-center text-lg font-bold outline-none transition-all"
-              style={{
-                borderColor: inputBorder,
-                background: inputBg,
-                color: inputColor,
-                boxShadow: 'inset 0 2px 4px rgba(0,0,0,0.02)'
-              }}
-            />
-            <button
-              type="submit"
-              disabled={chosen !== null || !inputValue.trim()}
-              className="w-full py-4 rounded-xl text-sm font-black tracking-wide transition-all disabled:opacity-50"
-              style={{ background: LIB.gold, color: LIB.ink }}
-            >
-              제출하기
-            </button>
-            {chosen !== null && chosen.trim().toLowerCase() !== question.answer.toLowerCase() && (
-              <div className="mt-2 text-center font-bold text-sm" style={{ color: '#991b1b' }}>
-                정답: {question.answer}
-              </div>
-            )}
-          </form>
-        ) : (
-          <div className="space-y-3">
-            {question.choices.map((choice, i) => {
-              let bg = 'white', borderColor = LIB.shelfLine, textColor = LIB.ink
-
         {hasBookAnim && (
           <BookStackAnimation count={stackedBooks} isToppling={isToppling} total={total} />
         )}
-
 
         <div className="space-y-3">
           {question.choices.map((choice, i) => {
@@ -618,24 +463,13 @@ function QuizView({ question, current, total, chosen, onChoose, activeLevel, has
               </button>
             )
           })}
-          </div>
-        )}
+        </div>
       </div>
     </div>
   )
 }
 
 // ── 결과 화면 ─────────────────────────────────────────────────────
-
- feat/quiz-level
-function ResultView({ answers, total, saving, onRetry, level }) {
-  const correctCount = answers.filter(a => a.correct).length
-  const score = Math.round((correctCount / total) * 100)
-
-  const grade = score >= 90 ? { label: '완벽해요!', icon: <Trophy size={48} strokeWidth={1.2} style={{ color: LIB.gold }} />, color: LIB.gold }
-    : score >= 70 ? { label: '잘했어요!', icon: <BookOpen size={48} strokeWidth={1.2} style={{ color: '#4ade80' }} />, color: '#4ade80' }
-      : score >= 50 ? { label: '조금 더!', icon: <BookMarked size={48} strokeWidth={1.2} style={{ color: LIB.woodLight }} />, color: LIB.woodLight }
-        : { label: '다시 도전!', icon: <Dumbbell size={48} strokeWidth={1.2} style={{ color: LIB.deepRed }} />, color: LIB.deepRed }
 
 function ResultView({ answers, total, saving, onRetry, quizResult, activeLevel, isUserWb }) {
   const correctCount  = answers.filter(a => a.correct).length
@@ -656,9 +490,6 @@ function ResultView({ answers, total, saving, onRetry, quizResult, activeLevel, 
           <div className="px-8 py-6 text-center relative" style={{ background: `linear-gradient(135deg, ${LIB.wood} 0%, ${LIB.woodLight} 100%)` }}>
             <div className="flex justify-center mb-2">{grade.icon}</div>
             <p className="text-xs font-bold uppercase tracking-widest mb-1" style={{ color: LIB.shelfLine }}>
- feat/quiz-level
-              Lv {level} 테스트 결과
-
               Lv {activeLevel} 테스트 결과
             </p>
             <p className="text-5xl font-black" style={{ color: LIB.gold }}>{score}</p>
@@ -733,22 +564,7 @@ function ResultView({ answers, total, saving, onRetry, quizResult, activeLevel, 
 // ── 메인 ─────────────────────────────────────────────────────────
 
 export default function QuizPage() {
- feat/quiz-level
-  const { user } = useAuth()
-  const [step, setStep] = useState('select')
-  const [wordbooks, setWordbooks] = useState([])
-  const [selectedWb, setSelectedWb] = useState(null)
-  const [questions, setQuestions] = useState([])
-  const [current, setCurrent] = useState(0)
-  const [answers, setAnswers] = useState([])
-  const [chosen, setChosen] = useState(null)
-  const [saving, setSaving] = useState(false)
-  const [activeLevel, setActiveLevel] = useState(1)
-  const [wbWords, setWbWords] = useState([])
-  const [levelStats, setLevelStats] = useState({ 1: 0, 2: 0, 3: 0, 4: 0 })
-  const [loadingWb, setLoadingWb] = useState(false)
-
-  const { user, profile } = useAuth()
+  const { user, profile }           = useAuth()
   const reward = useReward()
 
   const [step, setStep]             = useState('select')
@@ -764,7 +580,7 @@ export default function QuizPage() {
   const [noWordsModal, setNoWordsModal] = useState(false)
   const [wbLevelCounts, setWbLevelCounts] = useState({})
   const [stackedBooks, setStackedBooks] = useState(0)
-  const [isToppling, setIsToppling]   = useState(false)
+  const [isToppling,  setIsToppling]   = useState(false)
   const [wordCountMode, setWordCountMode] = useState('all')
 
   const achievementWbId = selectedWb?.id ?? null
@@ -781,78 +597,6 @@ export default function QuizPage() {
     ])
     const wbs = [
       ...(official ?? []).map(w => ({ ...w, type: 'official' })),
- feat/quiz-level
-      ...(mine ?? []).map(w => ({ ...w, type: 'user' })),
-    ])
-  }
-
-  const handleSelectWb = async (wb) => {
-    setSelectedWb(wb)
-    setWbWords([])
-    setLevelStats({ 1: 0, 2: 0, 3: 0, 4: 0 })
-    if (!wb) return
-
-    setLoadingWb(true)
-    let allWords = []
-    
-    if (wb.type === 'official') {
-      const { data: words } = await supabase
-        .from('official_words')
-        .select('id, english, major_meaning, major_example, general_example')
-        .eq('wordbook_id', wb.id)
-      
-      const { data: progress, error: progErr } = await supabase
-        .from('user_official_word_progress')
-        .select('word_id, word_level')
-        .eq('user_id', user.id)
-      
-      if (progErr) {
-        console.error('Progress fetch error:', progErr)
-      }
-      
-      const progressMap = {}
-      progress?.forEach(p => { progressMap[p.word_id] = p.word_level })
-      
-      allWords = (words || []).map(w => ({
-        ...w,
-        word_level: progressMap[w.id] || 1
-      }))
-    } else {
-      const { data } = await supabase
-        .from('user_words')
-        .select('id, english, major_meaning, major_example, general_example, word_level')
-        .eq('wordbook_id', wb.id)
-      allWords = data || []
-    }
-    
-    const stats = { 1: 0, 2: 0, 3: 0, 4: 0 }
-    allWords.forEach(w => {
-      const lv = w.word_level || 1
-      if (stats[lv] !== undefined) stats[lv]++
-    })
-    
-    setWbWords(allWords)
-    setLevelStats(stats)
-    setLoadingWb(false)
-  }
-
-  const startQuiz = async (level) => {
-    if (!selectedWb || loadingWb) return
-    setActiveLevel(level)
-    
-    if (wbWords.length < 4) {
-      alert('단어장에 단어가 4개 이상 필요합니다.')
-      return
-    }
-
-    const targetWords = wbWords.filter(w => (w.word_level || 1) === level)
-    if (targetWords.length === 0) {
-      alert(`해당 레벨(${level})에 도달한 단어가 없습니다.`)
-      return
-    }
-
-    setQuestions(buildQuestions(targetWords, wbWords, level))
-
       ...(mine    ?? []).map(w => ({ ...w, type: 'user' })),
     ]
     setWordbooks(wbs)
@@ -931,15 +675,12 @@ export default function QuizPage() {
       const { data } = await q
       allWords = data
     }
-    if (!allWords || allWords.length < 4) {
-      alert('퀴즈를 위해 단어가 4개 이상 필요합니다.')
-      return
-    }
-
-    setQuizResult(null)
-    setQuestions(buildQuestions(allWords))
     if (!allWords || allWords.length === 0) {
       setNoWordsModal(true)
+      return
+    }
+    if (allWords.length < 4) {
+      alert('퀴즈를 위해 단어가 4개 이상 필요합니다.')
       return
     }
 
@@ -960,11 +701,8 @@ export default function QuizPage() {
   const handleChoose = (choice) => {
     if (chosen !== null) return
     setChosen(choice)
-    const currentQ = questions[current]
-    const isCorrect = typeof choice === 'string' 
-      ? choice.trim().toLowerCase() === currentQ.answer.toLowerCase()
-      : choice === currentQ.answer
-    const nextAnswers = [...answers, { wordId: currentQ.id, chosen: choice, correct: isCorrect }]
+    const isCorrect   = choice === questions[current].answer
+    const nextAnswers = [...answers, { chosen: choice, correct: isCorrect }]
 
     if (hasBookAnim) {
       if (isCorrect) {
@@ -989,37 +727,12 @@ export default function QuizPage() {
   const finishQuiz = async (finalAnswers) => {
     setSaving(true)
     setAnswers(finalAnswers)
- feat/quiz-level
-    const correctAnswers = finalAnswers.filter(a => a.correct)
-    const correctCount = correctAnswers.length
-    const score = Math.round((correctCount / questions.length) * 100)
-    const correctWordIds = correctAnswers.map(a => a.wordId)
-
-    const { error } = await supabase.rpc('save_quiz_result', {
-      p_user_id: user.id,
-      p_wordbook_id: selectedWb.id,
-      p_wordbook_type: selectedWb.type,
-      p_score: score,
-      p_total: questions.length,
-      p_correct: correctCount,
-      p_level: activeLevel,
-      p_correct_word_ids: correctWordIds,
-    })
-
-    if (error) {
-      console.error('Quiz save error:', error)
-      alert('결과 저장 중 오류가 발생했습니다. 백엔드 연결이나 마이그레이션 적용 여부를 확인해주세요.\n' + error.message)
-    }
-
 
     const correctCount   = finalAnswers.filter(a => a.correct).length
     const score          = Math.round((correctCount / questions.length) * 100)
     const correctWordIds = finalAnswers
       .map((a, i) => a.correct ? questions[i].id : null)
       .filter(Boolean)
-
-    await endSession(sessionIdRef.current)
-    sessionIdRef.current = null
 
     const params = {
       p_user_id:          user.id,
@@ -1042,14 +755,6 @@ export default function QuizPage() {
   }
 
   if (step === 'select') return (
- feat/quiz-level
-    <SelectView wordbooks={wordbooks} selectedWb={selectedWb} onSelect={handleSelectWb} onStart={startQuiz} levelStats={levelStats} loadingWb={loadingWb} />
-  )
-  if (step === 'quiz') return (
-    <QuizView question={questions[current]} current={current} total={questions.length} chosen={chosen} onChoose={handleChoose} level={activeLevel} />
-  )
-  if (step === 'result') return (
-    <ResultView answers={answers} total={questions.length} saving={saving} onRetry={() => { setStep('select'); setSelectedWb(null) }} level={activeLevel} />
     <>
       <SelectView
         wordbooks={wordbooks}
@@ -1061,7 +766,6 @@ export default function QuizPage() {
         progress={progress}
         isReviewMode={isReviewMode}
         wbLevelCounts={wbLevelCounts}
-
         wordCountMode={wordCountMode}
         onWordCountModeChange={setWordCountMode}
       />
